repair memory display

+++ src/macsniff.cpp
@@ -27,7 +27,7 @@ bool mac_add(uint8_t *paddr, int8_t rssi, bool sniff_type) {
     bool added = false;
     uint32_t addr2int, vendor2int;          // temporary buffer for MAC and Vendor OUI
 	uint16_t hashedmac;                     // temporary buffer for generated hash value
-    float memlevel;                         // % of used heap mem
+    uint8_t memlevel;                         // % of used heap mem
 
     // only last 3 MAC Address bytes are used for MAC address anonymization
     // but since it's uint32 we take 4 bytes to avoid 1st value to be 0
@@ -53,7 +53,7 @@ bool mac_add(uint8_t *paddr, int8_t rssi, bool sniff_type) {
     if (added) {
         // Display heap memory left
         memlevel = ESP.getFreeHeap() / heapmem * 100;
-        sprintf(display_mem, "%d%%", memlevel);
+        sprintf(display_mem, "%i%%", memlevel);
         // increment counter and one blink led
         if (sniff_type == MAC_SNIFF_WIFI ) {
             macs_wifi++; // increment Wifi MACs counter
